refactor: Move a scenario from authentication to enter time
