+++ src/test/kotlin/coverosR3z/TestHelpers.kt
@@ -1,4 +1,7 @@
 package coverosR3z
+import coverosR3z.authentication.AuthenticationPersistence
+import coverosR3z.authentication.AuthenticationUtilities
+import coverosR3z.authentication.CurrentUserAccessor
 import coverosR3z.authentication.FakeCurrentUserAccessor
 import coverosR3z.domainobjects.*
 import coverosR3z.persistence.PureMemoryDatabase
@@ -6,6 +9,7 @@ import coverosR3z.timerecording.ITimeEntryPersistence
 import coverosR3z.timerecording.TimeEntryPersistence
 import coverosR3z.timerecording.TimeRecordingUtilities
 import kotlinx.serialization.json.Json
+import org.junit.Assert
 
 /**
  * a test helper method to create a [TimeEntry]
@@ -71,4 +75,28 @@ val jsonSerialzationWithPrettyPrint : Json = Json{prettyPrint = true; allowStruc
 fun createTimeRecordingUtility(): TimeRecordingUtilities {
         val timeEntryPersistence : ITimeEntryPersistence = TimeEntryPersistence(PureMemoryDatabase())
         return TimeRecordingUtilities(timeEntryPersistence, FakeCurrentUserAccessor())
+}
+
+/**
+ * Create an employee, "Alice", register a user for her, create a project
+ */
+fun initializeAUserAndLogin(cua : CurrentUserAccessor) : Pair<TimeRecordingUtilities, Employee>{
+        val pmd = PureMemoryDatabase()
+        val authPersistence = AuthenticationPersistence(pmd)
+        val au = AuthenticationUtilities(authPersistence, cua)
+
+        val tru = TimeRecordingUtilities(TimeEntryPersistence(pmd), cua)
+        val alice = tru.createEmployee(EmployeeName("Alice"))
+
+        au.register("alice", DEFAULT_PASSWORD, alice.id)
+        au.login("alice", DEFAULT_PASSWORD)
+
+        // Perform some quick checks
+        Assert.assertEquals("Auth persistence and user persistence must agree",
+                authPersistence.getUser(UserName("alice")), cua.get())
+        Assert.assertTrue("Registration must have succeeded", au.isUserRegistered("alice"))
+
+        tru.createProject(DEFAULT_PROJECT_NAME)
+
+        return Pair(tru, alice)
 }
